+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import {Component} from '@angular/core';
-import {FeEngineerFormComponent} from './fe-engineer-form/fe-engineer-form.component';
+import {FeEngineerFormComponent} from './features/fe-engineer-form/fe-engineer-form.component';
 
 @Component({
   selector: 'app-root',
