+++ src/app/features/fe-engineer-form/fe-engineer-form.component.html
@@ -1,6 +1,6 @@
 <div class="container-sm">
   <div class="">
-    <form (ngSubmit)="onSubmit()" [formGroup]="form" class="row mt-3 g-3">
+    <form [formGroup]="form" class="row mt-3 g-3">
       <!-- header -->
       <mat-toolbar>
         <span>Frontend Engineer Form</span>
@@ -65,7 +65,7 @@
 
       <!-- hobbies Input -->
       <div formArrayName="hobbies">
-        @for (hobby of hobbies.controls; track $index) {
+        @for (hobby of form.value.hobbies; track $index) {
           <div [formGroupName]="$index" class="row g-3">
             <mat-form-field class="col-md-4">
               <mat-label>Hobby Name</mat-label>
@@ -84,13 +84,13 @@
         }
 
         <button (click)="addHobby()" mat-raised-button type="button">Add Hobby</button>
-        @if (hobbiesIsInvalid) {
+        @if (hobbiesIsInvalid()) {
           <mat-error> Please add at least one hobby.</mat-error>
         }
       </div>
 
       <!-- submit button -->
-      <button [disabled]="form.invalid" class="w-25" mat-stroked-button>Send</button>
+      <button [disabled]="form.invalid" class="w-25" mat-stroked-button (click)="onSubmit()">Send</button>
     </form>
   </div>
 </div>
